-FIX: Renamed Functions to fit Saschas Request -DEL: Removed Test-Functions; they wer just testing "ThemSelf", no real function of the API. -FIX: Optimized JavaDoc
