+++ agrirouter-api-java-impl/src/main/java/com/dke/data/agrirouter/impl/onboard/OnboardingServiceImpl.java
@@ -87,12 +87,12 @@ public class OnboardingServiceImpl extends AbstractOnboardingService
   }
 
   @Override
-  public String getLastErrorAsString() {
+  public String getLastError() {
     return this.lastError;
   }
 
   @Override
-  public Optional<OnboardingError> getLastError() {
+  public Optional<OnboardingError> getLastOnboardingError() {
     if (this.lastError == null || this.lastError.equals("")) {
       return Optional.empty();
